refactor suggestions into seperate file, fix spacing inside tagButton
