+++ packages/ndla-ui/src/TagSelector/SuggestionInput.tsx
@@ -12,19 +12,15 @@ import { useTranslation } from 'react-i18next';
 import styled from '@emotion/styled';
 import Button, { IconButtonDualStates } from '@ndla/button';
 import { ChevronDown, ChevronUp } from '@ndla/icons/common';
-import { Cross } from '@ndla/icons/action';
-import { Check } from '@ndla/icons/editor';
-import { spacing, spacingUnit, colors, misc, animations, fonts, shadows } from '@ndla/core';
+import { Cross as CrossRaw } from '@ndla/icons/action';
+import { spacing, colors, misc, animations, fonts } from '@ndla/core';
 import Tooltip from '@ndla/tooltip';
 import { uuid } from '@ndla/util';
+import Suggestions from './Suggestions';
 import type { TagProp } from './TagSelector';
 
-const ABSOLUTE_DROPDOWN_MAXHEIGHT = '360px';
-
-const CheckedIcon = styled(Check)`
-  width: ${spacingUnit}px;
-  height: ${spacingUnit}px;
-  fill: ${colors.brand.light};
+const Cross = styled(CrossRaw)`
+  margin-left: ${spacing.xxsmall};
 `;
 
 const SuggestionInputContainer = styled.div`
@@ -57,70 +53,6 @@ const StyledInputWrapper = styled.div`
 const CombinedInputAndDropdownWrapper = styled.div`
   display: flex;
   flex-grow: 1;
-`;
-
-interface SuggestionsWrapperProps {
-  dropdownMaxHeight: string;
-  inline?: boolean;
-}
-
-const SuggestionsWrapper = styled.div`
-  position: relative;
-`;
-
-const Suggestions = styled.div<SuggestionsWrapperProps>`
-  position: ${({ inline }) => (inline ? 'static' : 'absolute')};
-  z-index: 99999;
-  right: 0;
-  left: 0;
-  box-shadow: ${shadows.levitate1};
-  margin: 0 ${spacing.small};
-  padding: ${spacing.small} 0;
-  overflow-y: scroll;
-  scroll-behavior: smooth;
-  max-height: min(${({ dropdownMaxHeight }) => dropdownMaxHeight}, ${ABSOLUTE_DROPDOWN_MAXHEIGHT});
-  border-radius: ${misc.borderRadius};
-  background: ${colors.white};
-  ${animations.fadeIn(animations.durations.fast)}
-  > div {
-    opacity: 0;
-    ${animations.fadeInBottom()}
-    animation-delay: ${animations.durations.fast};
-    animation-fill-mode: forwards;
-    display: flex;
-    flex-direction: column;
-  }
-`;
-
-interface SuggestionButtonProps {
-  isHighlighted: boolean;
-}
-
-const SuggestionButton = styled.button<SuggestionButtonProps>`
-  display: flex;
-  align-items: space-between;
-  justify-content: space-between;
-  border: 0;
-  outline: 0;
-  background: ${({ isHighlighted }) => (isHighlighted ? colors.brand.lighter : 'transparent')};
-  width: 100%;
-  padding: ${spacing.small};
-  ${fonts.sizes(18)};
-  transition: ${misc.transition.default};
-  &:not(:disabled) {
-    cursor: pointer;
-    color: ${colors.brand.primary};
-    &:hover {
-      background: ${colors.brand.lighter};
-    }
-  }
-  &:disabled {
-    &:hover {
-      svg {
-        fill: ${colors.brand.greyLight};
-      }
-    }
-  }
 `;
 
 interface SuggestionInputProps {
@@ -199,7 +131,6 @@ const SuggestionInput = ({
             size="small">
             {prefix}
             {name}
-            {` `}
             <Cross />
           </Button>
         ))}
@@ -283,31 +214,14 @@ const SuggestionInput = ({
       </StyledInputWrapper>
       <div id={suggestionIdRef.current} aria-live="polite">
         {(hasFocus || expanded) && suggestions.length > 0 ? (
-          <SuggestionsWrapper>
-            <Suggestions inline={inline} dropdownMaxHeight={dropdownMaxHeight}>
-              <div role="listbox">
-                {suggestions.map(({ id, name }, index: number) => {
-                  const alreadyAdded = hasBeenAdded(id);
-                  const selected = index === currentHighlightedIndex;
-                  return (
-                    <SuggestionButton
-                      data-suggestionbutton
-                      role="option"
-                      aria-selected={selected}
-                      disabled={alreadyAdded}
-                      isHighlighted={selected}
-                      onMouseDown={() => {
-                        onToggleTag(id);
-                      }}
-                      key={id}>
-                      <span>{name}</span>
-                      {alreadyAdded && <CheckedIcon />}
-                    </SuggestionButton>
-                  );
-                })}
-              </div>
-            </Suggestions>
-          </SuggestionsWrapper>
+          <Suggestions
+            inline={inline}
+            dropdownMaxHeight={dropdownMaxHeight}
+            suggestions={suggestions}
+            currentHighlightedIndex={currentHighlightedIndex}
+            onToggleTag={onToggleTag}
+            hasBeenAdded={hasBeenAdded}
+          />
         ) : null}
       </div>
     </SuggestionInputContainer>
